Partially change user follows
